+++ demo/main.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pyvista as pv
 
-import demo.quering
 from demo.utils import show_top_20_instances, show_heatmap
 from instance_masks_from_images import image_text
 
@@ -13,24 +12,27 @@ mesh = pv.read(r"C:\Users\sankl\Downloads\RA\RA_1M.ply")
 loaded_masks = np.load(r'C:\Users\sankl\Downloads\merged_masks.npz')
 loaded_text = np.load(r'C:\Users\sankl\Downloads\merged_embeddings.npz')
 
-
 QUERY = "green"
 
 mask_embeddings = {key: loaded_masks[key] for key in loaded_masks}
 
 mask_text_embeddings = {key: loaded_text[key].reshape(768, ) for key in loaded_text}
-model = image_text.load_image_text_model("google/siglip-base-patch16-224")
-query_embeddings = demo.quering.get_query_embedding(model, QUERY)
 
-scores = demo.quering.compute_cosine_similarity_scores(mask_text_embeddings, query_embeddings)
+model = image_text.load_image_text_model("google/siglip-base-patch16-224")
+query_embeddings = image_text.get_query_embedding(model, QUERY)
+
+scores = image_text.compute_cosine_similarity_scores(mask_text_embeddings, query_embeddings)
 
 # Create a plotter object
 plotter = pv.Plotter()
 plotter.add_text(f"Query: '{QUERY}'", font_size=14)
 
-# Add the point cloud or mesh to the plotter
+# Add the mesh to the plotter
 plotter.add_mesh(mesh, scalars='RGBA', rgb=True)
-#show_top_20_instances(plotter, mesh, scores, mask_embeddings)
+
+# Add masks as points on top
+# show_top_20_instances(plotter, mesh, scores, mask_embeddings)
 show_heatmap(plotter, mesh, scores, mask_embeddings)
+
 # Display the plot
 plotter.show()
